refactor: replace module `app.core` with module `app._core`
