+++ src/aria2_server/app/_core/static_files.py
